refactor: split docs markdown html renderer
